Switched uintptr to unsigned long long type
Previously, uintptr was indistinguishable from uint64, which
had various problems. Now it's a different type with the
same size. This enables proper type analysis.

+++ kernel/std.h
@@ -16,7 +16,7 @@ typedef signed short int16;
 typedef signed int int32;
 typedef signed long int64;
 
-typedef unsigned long uintptr;
+typedef unsigned long long uintptr;
 typedef float float32;
 typedef double float64;
